# packages/docs/openapi/data-engine.yaml
openapi: 3.0.0
info:
    title: Data Engine API
    description: |
        API for managing structured data resources including domains, URLs, calls, tags, documents, tickets, and chats.

        This API provides comprehensive CRUD operations for managing data across multiple domains with support for filtering, pagination, and advanced search capabilities.
    version: 1.0.0
    contact:
        name: AnswerAgentAI Support
        url: https://theanswer.ai
        email: support@theanswer.ai
servers:
    - url: https://prod.studio.theanswer.ai/api/v1/data-engine
      description: Production server
    - url: http://localhost:4000/api/v1/data-engine
      description: Local development server
...
paths:
...
    /tags:
        post:
            summary: Create a new tag
            operationId: createTag
            tags:
                - tags
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TagInput'
            responses:
                '201':
                    description: Tag created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tag'
        get:
            summary: List all tags
            operationId: listTags
            tags:
                - tags
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TagListResponse'
    /tags/hierarchy:
        get:
            summary: Get tag hierarchy
            description: Retrieve tags in hierarchical structure (parent tags with subcategories)
            operationId: getTagHierarchy
            tags:
                - tags
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties:
                                    $ref: '#/components/schemas/TagHierarchyItem'
    /tags/{id}:
        get:
            summary: Get tag by ID
            operationId: getTagById
            tags:
                - tags
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                  description: Tag ID
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tag'
        put:
            summary: Update tag
            operationId: updateTag
            tags:
                - tags
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TagUpdate'
            responses:
                '200':
                    description: Tag updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tag'
        delete:
            summary: Delete tag
            operationId: deleteTag
            tags:
                - tags
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Tag deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
    # ==================== DOCUMENTS ====================
...
    /tickets:
        post:
            summary: Create a new ticket
            operationId: createTicket
            tags:
                - tickets
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TicketInput'
            responses:
                '201':
                    description: Ticket created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Ticket'
        get:
            summary: List all tickets
            operationId: listTickets
            tags:
                - tickets
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/PageParam'
                - $ref: '#/components/parameters/PageSizeParam'
                - name: status
                  in: query
                  schema:
                      type: string
                - name: assignedTo
                  in: query
                  schema:
                      type: string
                - name: tags
                  in: query
                  description: Array of tag strings (JSON string)
                  schema:
                      type: string
                      example: '["urgent", "customer"]'
                - name: startDate
                  in: query
                  schema:
                      type: string
                      format: date-time
                - name: endDate
                  in: query
                  schema:
                      type: string
                      format: date-time
                - name: ticketType
                  in: query
                  schema:
                      type: string
                - name: escalated
                  in: query
                  schema:
                      type: boolean
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TicketListResponse'
    /tickets/{id}:
        get:
            summary: Get ticket by ID
            operationId: getTicketById
            tags:
                - tickets
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Ticket'
        put:
            summary: Update ticket
            operationId: updateTicket
            tags:
                - tickets
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TicketUpdate'
            responses:
                '200':
                    description: Ticket updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Ticket'
        delete:
            summary: Delete ticket
            operationId: deleteTicket
            tags:
                - tickets
            security:
                - bearerAuth: []
            parameters:
                - $ref: '#/components/parameters/IdParam'
            responses:
                '200':
                    description: Ticket deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
    # ==================== CHATS ====================
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: API Key
            description: |
                API key authentication. Include your API key in the Authorization header:
                `Authorization: Bearer YOUR_API_KEY`
    parameters:
        PageParam:
            name: page
            in: query
            description: Page number (0-indexed)
            schema:
                type: integer
                default: 0
                minimum: 0
        PageSizeParam:
            name: pageSize
            in: query
            description: Items per page
            schema:
                type: integer
                default: 10
                minimum: 1
                maximum: 100
...
        IdParam:
            name: id
            in: path
            required: true
            description: Resource ID (UUID)
            schema:
                type: string
                format: uuid
    schemas:
...
        Metadata:
            type: object
            description: Metadata object for tracking source system and user information
            properties:
                source_system:
                    type: string
                    example: theanswer
                source_organization_id:
                    type: string
                    format: uuid
                source_user_id:
                    type: string
                    format: uuid
                created_by:
                    type: string
                    format: email
                last_updated_by:
                    type: string
                    format: email
                last_updated_from:
                    type: string
                    example: theanswer
        DeleteResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: Resource deleted successfully
        PaginationInfo:
            type: object
            properties:
                total:
                    type: integer
                    example: 100
                page:
                    type: integer
                    example: 0
                pageSize:
                    type: integer
                    example: 10
        # Domain Schemas
...
        TagInput:
            type: object
            required:
                - slug
                - label
            properties:
                slug:
                    type: string
                    example: product-inquiry
                label:
                    type: string
                    example: Product Inquiry
                description:
                    type: string
                    example: Questions about products
                color:
                    type: string
                    example: '#FF5733'
                shade:
                    type: string
                    example: '#FF8C66'
                parent_id:
                    type: integer
                    nullable: true
                metadata:
                    $ref: '#/components/schemas/Metadata'
        Tag:
            allOf:
                - $ref: '#/components/schemas/TagInput'
                - type: object
                  properties:
                      id:
                          type: integer
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
        TagUpdate:
            type: object
            properties:
                label:
                    type: string
                description:
                    type: string
                color:
                    type: string
                shade:
                    type: string
                metadata:
                    $ref: '#/components/schemas/Metadata'
        TagListResponse:
            type: object
            properties:
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/Tag'
        TagHierarchyItem:
            type: object
            properties:
                label:
                    type: string
                description:
                    type: string
                color:
                    type: string
                slug:
                    type: string
                id:
                    type: integer
                subcategories:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/TagHierarchyItem'
        # Document Schemas
...
        TicketInput:
            type: object
            required:
                - title
                - description
            properties:
                title:
                    type: string
                    example: Ticket Title
                description:
                    type: string
                    example: Ticket description
                status:
                    type: string
                    enum: [open, in_progress, resolved, closed]
                    example: open
                priority:
                    type: string
                    enum: [low, medium, high, urgent]
                    example: high
                assigned_to:
                    type: string
                    format: email
                    example: user@example.com
                ticket_type:
                    type: string
                    example: email
                escalated:
                    type: boolean
                    example: false
                tags_array:
                    type: array
                    items:
                        type: string
                    example: [urgent, customer]
                metadata:
                    $ref: '#/components/schemas/Metadata'
        Ticket:
            allOf:
                - $ref: '#/components/schemas/TicketInput'
                - type: object
                  properties:
                      id:
                          type: string
                          format: uuid
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
        TicketUpdate:
            type: object
            properties:
                status:
                    type: string
                    enum: [open, in_progress, resolved, closed]
                priority:
                    type: string
                    enum: [low, medium, high, urgent]
                assigned_to:
                    type: string
                    format: email
                tags_array:
                    type: array
                    items:
                        type: string
                metadata:
                    $ref: '#/components/schemas/Metadata'
        TicketListResponse:
            allOf:
                - $ref: '#/components/schemas/PaginationInfo'
                - type: object
                  properties:
                      tickets:
                          type: array
                          items:
                              $ref: '#/components/schemas/Ticket'
        # Chat Schemas
...